move export.hpp into aux_ folder. put assert symbols in libtorrent namespace
