+++ include/libtorrent/Makefile.am
@@ -44,7 +44,6 @@ nobase_include_HEADERS = \
   enum_net.hpp                 \
   error.hpp                    \
   error_code.hpp               \
-  export.hpp                   \
   extensions.hpp               \
   file.hpp                     \
   file_pool.hpp                \
@@ -175,6 +174,7 @@ nobase_include_HEADERS = \
   aux_/dev_random.hpp               \
   aux_/deque.hpp                    \
   aux_/escape_string.hpp            \
+  aux_/export.hpp                   \
   aux_/io.hpp                       \
   aux_/listen_socket_handle.hpp     \
   aux_/max_path.hpp                 \
