+++ include/libtorrent/config.hpp
@@ -43,7 +43,7 @@ POSSIBILITY OF SUCH DAMAGE.
 
 // TODO: don't include that here. Make each header that use the export macros
 // include it instead. and move it to aux_
-#include "libtorrent/export.hpp"
+#include "libtorrent/aux_/export.hpp"
 
 #ifdef __linux__
 #include <linux/version.h> // for LINUX_VERSION_CODE and KERNEL_VERSION
