+++ include/libtorrent/ed25519.hpp
@@ -1,7 +1,7 @@
 #ifndef ED25519_HPP
 #define ED25519_HPP
 
-#include "libtorrent/export.hpp" // for TORRENT_EXPORT
+#include "libtorrent/aux_/export.hpp" // for TORRENT_EXPORT
 #include <stddef.h> // for size_t
 
 namespace libtorrent {
