+++ include/libtorrent/write_resume_data.hpp
@@ -36,7 +36,7 @@ POSSIBILITY OF SUCH DAMAGE.
 #include <vector>
 
 #include "libtorrent/fwd.hpp"
-#include "libtorrent/export.hpp"
+#include "libtorrent/aux_/export.hpp"
 #include "libtorrent/bencode.hpp"
 
 namespace libtorrent {
